requirements.txt
# Framework Web
Flask==2.3.3

# Base de datos PostgreSQL
psycopg2-binary==2.9.11  # Compatible con Python 3.13-3.14 [citation:5]

# Procesamiento de imágenes y video
opencv-python==4.12.0.88  # Soporte oficial para Python 3.13 [citation:2]
numpy==2.2.6  # Versión estable con soporte para Python 3.13 [citation:1][citation:8]
easyocr==1.7.2  # Última versión estable [citation:3][citation:9]

# Automatización web
playwright==1.49.0  # Requiere Python >=3.8 [citation:4]
requests==2.31.0

# Procesamiento de datos
pandas==2.2.3  # Compatible con numpy 2.x
openpyxl==3.1.2

# Utilidades
python-dotenv==1.0.0